maybe 찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐찐막

+++ src/main/java/com/ddiring/backend_market/trade/dto/VerifyTradeDto.java
@@ -16,10 +16,10 @@ public class VerifyTradeDto {
         private Long tradeId;
 
         @NotNull
-        private Integer buyId;
+        private Long buyId;
 
         @NotNull
-        private Integer sellId;
+        private Long sellId;
 
         @NotNull
         private Integer tradeAmount;
